[#234] RealmDB에 있는 사용자 정보를 처리하는 코드에 문서 작성
